challenge2/requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
coreapi==2.3.3
coreapi-cli==1.0.9
coreschema==0.0.4
Django==3.2.8
djangorestframework==3.12.4
gunicorn==20.1.0
idna==3.3
itypes==1.2.0
Jinja2==3.0.2
lxml==4.6.3
MarkupSafe==2.0.1
psycopg2-binary==2.9.2
python-dotenv==0.19.2
pytz==2021.3
PyYAML==6.0
requests==2.26.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.7
